requirements.txt
python-telegram-bot>=21.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
imapclient>=2.3.1
html2text>=2020.1.16
python-dateutil>=2.9.0

